refactor: tryCount 원시값 TryCount 클래스로 포장
